Fix linux UTs
