+++ unit_tests/fake/CMakeLists.txt
@@ -26,7 +26,6 @@ project(fake)
 #
 set(FAKE_THREAD_UTILS_SRC
   fake_error_manager.cpp
-  fake_page_server.cpp
   fake_system_parameter.cpp
   fake_thread_entry.cpp
   fake_thread_utils.cpp
