+++ unit_tests/log/CMakeLists.txt
@@ -153,6 +153,7 @@ add_dependencies(test_prior_sendrecv
 #
 set(TEST_LOG_PAGE_FETCHER_SOURCES
   ${STORAGE_DIR}/async_page_fetcher.cpp
+  dummy_page_server.cpp
   test_main_log_page_fetcher.cpp)
 
 SET_SOURCE_FILES_PROPERTIES(
@@ -212,7 +213,6 @@ add_dependencies(test_log_page_broker
 set(TEST_LOG_LSA_UTILS_SOURCES
   ${TRANSACTION_DIR}/log_lsa_utils.cpp
   ${BASE_DIR}/packer.cpp
-  ${TEST_FAKE_DIR}/fake_page_server.cpp
   test_main_log_lsa_utils.cpp)
 
 SET_SOURCE_FILES_PROPERTIES(
@@ -243,6 +243,7 @@ add_dependencies(test_log_lsa_utils
 
 set(TEST_DATA_PAGE_FETCHER_SOURCES
   ${STORAGE_DIR}/async_page_fetcher.cpp
+  dummy_page_server.cpp
   test_main_data_page_fetcher.cpp)
 
 SET_SOURCE_FILES_PROPERTIES(
